jsonrpc-types-proto: move proto definitions into its own crate
Closes: #8407
