+++ json-rpc/types/proto/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "jsonrpc-types-proto"
+version = "0.0.0"
+authors = ["Diem Association <opensource@diem.com>"]
+description = "Diem JSONRPC Proto Types"
+repository = "https://github.com/diem/diem"
+homepage = "https://diem.com"
+license = "Apache-2.0"
+publish = false
+edition = "2018"
+
+[dependencies]
+prost = "0.7.0"
+serde = { version = "1.0.124", default-features = false }
+serde_json = "1.0.64"
+
+diem-types = { path = "../../../types" }
+diem-json-rpc-types = { path = ".." }
+diem-workspace-hack = { path = "../../../common/workspace-hack" }
+
+[build-dependencies]
+prost-build = "0.7.0"
