+++ json-rpc/types/src/lib.rs
@@ -6,9 +6,6 @@ pub mod request;
 pub mod response;
 pub mod views;
 
-pub mod constants;
-pub mod proto;
-
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Copy, Debug, Deserialize, Serialize)]
@@ -66,37 +63,5 @@ impl Method {
             Method::GetTransactionsWithProofs => "get_transactions_with_proofs",
             Method::GetEventsWithProofs => "get_events_with_proofs",
         }
-    }
-}
-
-fn is_default<T: Default + PartialEq>(t: &T) -> bool {
-    t == &T::default()
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::{is_default, proto::types::Amount};
-
-    #[test]
-    fn test_is_default() {
-        let amount = Amount {
-            amount: 0,
-            currency: "CODE".to_string(),
-        };
-        assert!(is_default(&amount.amount));
-        assert!(!is_default(&amount.currency));
-
-        assert_eq!(
-            serde_json::json!({
-                "currency": "CODE"
-            }),
-            serde_json::to_value(&amount).unwrap()
-        );
-
-        let des_amount: Amount = serde_json::from_value(serde_json::json!({
-            "currency": "CODE"
-        }))
-        .unwrap();
-        assert_eq!(amount, des_amount);
     }
 }
